// src/services/api/generated/azeel.schemas.ts
export type UsersGetCustomerBundlesParams = {
  /**
   * @minimum 1
   * @maximum 100
   */
  per_page?: number;
  q?: string;
  sort?: string;
  direction?: UsersGetCustomerBundlesDirection;
};

export type UsersGetCustomerBundlesDirection =
  (typeof UsersGetCustomerBundlesDirection)[keyof typeof UsersGetCustomerBundlesDirection];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UsersGetCustomerBundlesDirection = {
  asc: "asc",
  desc: "desc",
} as const;